Inventario-Django/requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.6
django-model-utils==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.4
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0